+++ terraform/envs/shared/main.tf
@@ -21,8 +21,8 @@ module "github_oidc" {
   github_org  = var.github_org
   github_repo = var.github_repo
 
-  role_name   = var.github_actions_role_name
-  policy_name = var.github_actions_policy_name
+  role_name          = var.github_actions_role_name
+  inline_policy_name = var.github_actions_inline_policy_name
 
   oidc_thumbprint_list = var.oidc_thumbprint_list